+++ generators/java/sdk/src/main/java/com/fern/java/client/ClientPoetClassNameFactory.java
@@ -24,6 +24,14 @@ public final class ClientPoetClassNameFactory extends AbstractNonModelPoetClassN
                 errorDeclaration.getName().getName().getPascalCase().getSafeName());
     }
 
+    public ClassName getInputStreamRequestBodyClassName() {
+        return ClassName.get(getCorePackage(), "InputStreamRequestBody");
+    }
+
+    public ClassName getFileStreamClassName() {
+        return ClassName.get(getCorePackage(), "FileStream");
+    }
+
     public ClassName getRetryInterceptorClassName() {
         return ClassName.get(getCorePackage(), "RetryInterceptor");
     }