+++ packages/commons/core-utils/src/ObjectPropertiesVisitor.ts
@@ -4,7 +4,7 @@ export type ObjectPropertiesVisitor<T, R> = {
     [K in keyof T]-?: (value: T[K]) => R;
 };
 
-export async function visitObject<T extends Record<string, unknown>>(
+export async function visitObject<T extends object>(
     object: T,
     visitor: ObjectPropertiesVisitor<T, void | Promise<void>>
 ): Promise<void> {