+++ packages/commons/core-utils/src/index.ts
@@ -22,3 +22,5 @@ export type { Digit, Letter, LowercaseLetter, UppercaseLetter } from "./types";
 export { visitDiscriminatedUnion } from "./visitDiscriminatedUnion";
 export type { WithoutQuestionMarks } from "./withoutQuestionMarks";
 export { mergeWithOverrides } from "./mergeWithOverrides";
+export { titleCase } from "./titleCase";
+export { stripLeadingSlash } from "./stripLeadingSlash";