+++ packages/commons/core-utils/src/stripLeadingSlash.ts
@@ -1,0 +1,5 @@
+export function stripLeadingSlash(str: string): string;
+export function stripLeadingSlash(str: string | undefined): string | undefined;
+export function stripLeadingSlash(str: string | undefined): string | undefined {
+    return str?.replace(/^\/+/, "");
+}