+++ packages/commons/fs-utils/src/__test__/relativize.test.ts
@@ -6,7 +6,7 @@ describe("join", () => {
     it("simple", async () => {
         const from = AbsoluteFilePath.of("/path/to/fern");
         const to = AbsoluteFilePath.of("/path/to/fern/docs/markdown.md");
-        const path = await relativize(from, to);
+        const path = relativize(from, to);
         expect(path).toEqual(RelativeFilePath.of("docs/markdown.md"));
     });
 });